refactor: 중복 로또 번호 검증 주체 변경
- Numbers가 아닌 Lotto 객체에서 중복 로또 번호를 검증하도록 변경
- Numbers는 단순히 중복된 번호가 존재하는지 여부를 return 하는 기능만 제공하고 Lotto가 해당 기능을 이용하여 검증을 수행함
