+++ src/main/java/step3/domain/Lotto.java
@@ -9,6 +9,7 @@ public class Lotto {
     }
 
     private Lotto(Numbers numbers) {
+        validate(numbers);
         this.numbers = numbers;
     }
 
@@ -18,6 +19,12 @@ public class Lotto {
 
     public int getMatchCount(Numbers numbers) {
         return this.numbers.match(numbers);
+    }
+
+    private void validate(Numbers numbers) {
+        if (numbers.isDuplicated()) {
+            throw new IllegalArgumentException("Duplicate numbers cannot input.");
+        }
     }
 
     @Override
